Frame_Change Hack
The current problem with the frame change lagging behind a frame is from the current dependency graph. The shape keys on the objects don't get updated so they look as if they are lagging behind a frame. Without this hack you can manually update it by changing the frame and mute then unmute the shape key and you will see the shape key change. This is a temporary hack that forces the shape key to update and display the correct deformations.

When the new dependency graph is on by default in blender this hack will not be needed.